requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
google-genai==1.10.0
google-cloud-firestore>=2.11.0
google-cloud-logging>=3.6.0
slowapi>=0.1.8
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
httpx>=0.24.1
